Move IsExecutable check into Helpers library.
