+++ Lib/Collectors/LinuxFileSystemUtils.cs
@@ -13,72 +13,6 @@ namespace AttackSurfaceAnalyzer.Collectors
 {
     public class LinuxFileSystemUtils
     {
-        private static readonly List<string> MachMagicNumbers = new List<string>()
-        {
-            Helpers.HexStringToAscii("FEEDFACE"),
-            Helpers.HexStringToAscii("FEEDFACF"),
-            Helpers.HexStringToAscii("CEFAEDFE"),
-            Helpers.HexStringToAscii("CFFAEDFE"),
-            Helpers.HexStringToAscii("CAFEBEBE")
-        };
-
-        public static bool IsExecutable(string Path)
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            {
-                try
-                {
-                    var fourBytes = new byte[4];
-                    using (var fileStream = File.Open(Path, FileMode.Open))
-                    {
-                        fileStream.Read(fourBytes, 0, 4);
-                    }
-                    // ELF Format magic number " ELF"
-                    return (Encoding.ASCII.GetString(fourBytes) == Helpers.HexStringToAscii("7F454C46"));
-                }
-                catch (UnauthorizedAccessException)
-                {
-                    return false;
-                }
-                catch (IOException)
-                {
-                    return false;
-                }
-                catch (Exception e)
-                {
-                    Logger.DebugException(e);
-                    return false;
-                }
-            }
-            else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-            {
-                try
-                {
-                    var fourBytes = new byte[4];
-                    using (var fileStream = File.Open(Path, FileMode.Open))
-                    {
-                        fileStream.Read(fourBytes, 0, 4);
-                    }
-                    // Mach-o format magic numbers
-                    return MachMagicNumbers.Contains(Encoding.ASCII.GetString(fourBytes));
-                }
-                catch (UnauthorizedAccessException)
-                {
-                    return false;
-                }
-                catch (IOException)
-                {
-                    return false;
-                }
-                catch (Exception e)
-                {
-                    Logger.DebugException(e);
-                    return false;
-                }
-            }
-            return false;
-        }
-
         protected internal static string GetFilePermissions(FileSystemInfo fileInfo)
         {
             var filename = fileInfo.FullName;
